Split store tests
